+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/envs/base_env_cfg.py
@@ -12,6 +12,7 @@ configuring the environment instances, viewer settings, and simulation parameter
 from __future__ import annotations
 
 from dataclasses import MISSING
+from typing import Literal
 
 import omni.isaac.orbit.envs.mdp as mdp
 from omni.isaac.orbit.managers import RandomizationTermCfg as RandTerm
@@ -28,15 +29,40 @@ class ViewerCfg:
 
     eye: tuple[float, float, float] = (7.5, 7.5, 7.5)
     """Initial camera position (in m). Default is (7.5, 7.5, 7.5)."""
+
     lookat: tuple[float, float, float] = (0.0, 0.0, 0.0)
     """Initial camera target position (in m). Default is (0.0, 0.0, 0.0)."""
+
     cam_prim_path: str = "/OmniverseKit_Persp"
-    """The camera prim path to record images from. Default is "/OmniverseKit_Persp", which is the
-    default camera in the default viewport.
+    """The camera prim path to record images from. Default is "/OmniverseKit_Persp",
+    which is the default camera in the viewport.
     """
+
     resolution: tuple[int, int] = (1280, 720)
     """The resolution (width, height) of the camera specified using :attr:`cam_prim_path`.
     Default is (1280, 720).
+    """
+
+    origin_type: Literal["world", "env", "asset_root"] = "world"
+    """The frame in which the camera position (eye) and target (lookat) are defined in. Default is "world".
+
+    Available options are:
+
+    * ``"world"``: The origin of the world.
+    * ``"env"``: The origin of the environment defined by :attr:`env_index`.
+    * ``"asset_root"``: The center of the asset defined by :attr:`asset_name` in environment :attr:`env_index`.
+    """
+
+    env_index: int = 0
+    """The environment index for frame origin. Default is 0.
+
+    This quantity is only effective if :attr:`origin` is set to "env" or "asset_root".
+    """
+
+    asset_name: str | None = None
+    """The asset name in the interactive scene for the frame origin. Default is None.
+
+    This quantity is only effective if :attr:`origin` is set to "asset_root".
     """
 
 
@@ -62,7 +88,7 @@ class BaseEnvCfg:
     """Physics simulation configuration. Default is SimulationCfg()."""
     # ui settings
     ui_window_class_type: type | None = BaseEnvWindow
-    """The class type of the UI window. Defaults to None.
+    """The class type of the UI window. Default is None.
 
     If None, then no UI window is created.
 
@@ -84,5 +110,5 @@ class BaseEnvCfg:
     actions: object = MISSING
     """Action space settings."""
     randomization: object = DefaultRandomizationManagerCfg()
-    """Randomization settings. Defaults to the default randomization manager, which resets
+    """Randomization settings. Default is the default randomization manager, which resets
     the scene to its default state."""