+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/envs/ui/__init__.py
@@ -12,3 +12,4 @@ toggling different debug visualization tools, and other user-defined functionali
 
 from .base_env_window import BaseEnvWindow
 from .rl_task_env_window import RLTaskEnvWindow
+from .viewport_camera_controller import ViewportCameraController